Extract UI panel's pant menu listener registration.
